+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 * In general follow (https://docs.npmjs.com/getting-started/semantic-versioning) versioning.
 
 ## <next>
-
+* Added option selectComponentTranslations to customize react-selects default placeholder and noResultsTexts. By default uses the Grids translations.
 ## 5.1.8
 * Upgraded `@opuscapita/react-select-order-list` to the latest version.